+++ src/planner/apps/partner-in-crime/routes/plan.route.ts
@@ -2,8 +2,9 @@
 import { Request, Response } from "express";
 import { container } from "tsyringe";
 import { CreatePlanMessage } from "../../../controllers/plan/CreatePlanController";
-import { UserActor } from "../../../views/user-actor/UserActor";
 import { Identifier } from "../../../models/Identifier";
+import { User } from "../../../models/user/User";
+import { UserActor } from "../../../views/user-actor/UserActor";
 
 export const register = (app: any) => {
   const view = container.resolve(UserActor);
@@ -24,13 +25,7 @@ export const register = (app: any) => {
 
   app.post("/plans", async (req: Request, res: Response) => {
     try {
-      // TODO: get user id from the jwt in the Authorization Bearer header
-      // const user = await view.getUser(req.header.get("Authorization Bearer"));
-      const user = await view.getUser("626d3f7b7aa88b9339627665");
-
-      if (!user) {
-        return res.status(404).json();
-      }
+      const user: User = req["user"]
 
       const createPlanMessage: CreatePlanMessage = {
         owner: user.toPrimitives(),
@@ -49,7 +44,7 @@ export const register = (app: any) => {
         return res.status(500).json({ message: "Error creating plan" });
       }
 
-      return res.status(200).json({ success: true, planId: planId.toString() });
+      return res.status(201).json({ success: true, planId: planId.toString() });
     } catch (error) {
       console.error(error);
       res.status(500).send({ message: "internal-error" });
@@ -60,15 +55,7 @@ export const register = (app: any) => {
     "/plans/:planId/post-message",
     async (req: Request, res: Response) => {
       try {
-        // TODO: get user id from the jwt in the Authorization Bearer header
-        // const user = await view.getUser(req.header.get("Authorization Bearer"));
-        const user = await view.getUser("626d3f7b7aa88b9339627665");
-
-        if (!user) {
-          return res
-            .status(404)
-            .json({ message: "Requested plan no longer has an owner" });
-        }
+        const user: User = req["user"]
 
         const planId = req.params.planId as string;
         const content = req.body.content as string;
@@ -91,15 +78,7 @@ export const register = (app: any) => {
     "/plans/:planId/join-request",
     async (req: Request, res: Response) => {
       try {
-        // TODO: get user id from the jwt in the Authorization Bearer header
-        // const user = await view.getUser(req.header.get("Authorization Bearer"));
-        const user = await view.getUser("626d41aaf7dd454d1de11ffa");
-
-        if (!user) {
-          return res
-            .status(404)
-            .json({ message: "Requested plan no longer has an owner" });
-        }
+        const user: User = req["user"]
 
         const planId = req.params.planId as string;
 
@@ -133,15 +112,7 @@ export const register = (app: any) => {
     "/plans/:planId/join-requests",
     async (req: Request, res: Response) => {
       try {
-        // TODO: get user id from the jwt in the Authorization Bearer header
-        // const user = await view.getUser(req.header.get("Authorization Bearer"));
-        const user = await view.getUser("626d3f7b7aa88b9339627665");
-
-        if (!user) {
-          return res
-            .status(404)
-            .json({ message: "Requested plan no longer has an owner" });
-        }
+        const user: User = req["user"]
 
         const planId = req.params.planId as string;
         const joinPlanRequests = await view.getJoinPlanRequests();