+++ src/planner/apps/partner-in-crime/server.ts
@@ -4,6 +4,8 @@ import Router from "express-promise-router";
 import * as http from "http";
 import cors from "cors";
 import { registerRoutes } from "./routes";
+import { authenticateTokenMiddleware } from "./middlewares/VerifyToken";
+import { USER_PATH } from "./routes/user.route";
 
 export class Server {
   private express: express.Express;
@@ -15,6 +17,7 @@ export class Server {
     this.express = express();
     this.express.use(bodyParser.json());
     this.express.use(cors());
+    this.express.use(this.unless(USER_PATH.LOGIN, authenticateTokenMiddleware))
     const router = Router();
     this.express.use(router);
     registerRoutes(router);
@@ -25,8 +28,7 @@ export class Server {
       (resolve) => (this.httpServer = this.express.listen(this.port, resolve))
     );
     console.info(
-      `  HTTP App is running at http://localhost:${
-        this.port
+      `  HTTP App is running at http://localhost:${this.port
       } in ${this.express.get("env")} mode`
     );
     console.info("  Press CTRL-C to stop\n");
@@ -46,4 +48,14 @@ export class Server {
       return resolve();
     });
   }
+
+  private unless = (path, middleware) => {
+    return (req, res, next) => {
+      if (path === req.path) {
+        return next();
+      } else {
+        return middleware(req, res, next);
+      }
+    };
+  };
 }